Simplify types

+++ src/AudioCore/ADSR.h
@@ -22,45 +22,45 @@ public:
 	~ADSR();
 	void KeyOn();
 	void KeyOff();
-	void SetAttackRate( double rate );
-	void SetAttackTarget( double target );
-	void SetDecayRate( double rate );
-	void SetSustainLevel( double level );
-	void SetReleaseRate( double rate );
-	void SetAttackTime( double time );
-	void SetDecayTime( double time );
-	void SetReleaseTime( double time );
-	void SetAllTimes( double aTime, double dTime, double sLevel, double rTime );
-	void SetTarget( double target );
+	void SetAttackRate( float rate );
+	void SetAttackTarget( float target );
+	void SetDecayRate( float rate );
+	void SetSustainLevel( float level );
+	void SetReleaseRate( float rate );
+	void SetAttackTime( float time );
+	void SetDecayTime( float time );
+	void SetReleaseTime( float time );
+	void SetAllTimes( float aTime, float dTime, float sLevel, float rTime );
+	void SetTarget( float target );
 	
 	int GetState() const {
 		return state_;
 	};
 
-	void SetValue( double value );
+	void SetValue( float value );
 
-	double GetLastOut() const {
+	float GetLastOut() const {
 		return last_frame_[0];
 	};
 
-	double Tick();
-	AudioFrames& Tick( AudioFrames& frames, unsigned int channel = 0 );
+	float Tick();
+	AudioFrames& Tick( AudioFrames& frames, int channel = 0 );
 
 protected:
 
-	void SampleRateChanged( double new_rate, double old_rate );
+	void SampleRateChanged( float new_rate, float old_rate );
 
 	int state_;
-	double value_;
-	double target_;
-	double attack_rate_;
-	double decay_rate_;
-	double release_rate_;
-	double release_time_;
-	double sustain_level_;
+	float value_;
+	float target_;
+	float attack_rate_;
+	float decay_rate_;
+	float release_rate_;
+	float release_time_;
+	float sustain_level_;
 };
 
-inline double ADSR::Tick() {
+inline float ADSR::Tick() {
 	switch ( state_ ) {
 	case ATTACK:
 		value_ += attack_rate_;
@@ -98,8 +98,8 @@ inline double ADSR::Tick() {
 	case RELEASE:
 		value_ -= release_rate_;
 
-		if ( value_ <= 0.0 ) {
-			value_ = 0.0;
+		if ( value_ <= 0.0f ) {
+			value_ = 0.0f;
 			state_ = IDLE;
 		}
 
@@ -109,7 +109,7 @@ inline double ADSR::Tick() {
 	return value_;
 }
 
-inline AudioFrames& ADSR::Tick( AudioFrames& frames, unsigned int channel ) {
+inline AudioFrames& ADSR::Tick( AudioFrames& frames, int channel ) {
 	#if defined(flagDEBUG)
 
 	if ( channel >= frames.GetChannelCount() ) {
@@ -118,10 +118,10 @@ inline AudioFrames& ADSR::Tick( AudioFrames& frames, unsigned int channel ) {
 	}
 
 	#endif
-	double* samples = &frames[channel];
-	unsigned int step = frames.GetChannelCount();
+	float* samples = &frames[channel];
+	int step = frames.GetChannelCount();
 
-	for ( unsigned int i = 0; i < frames.GetFrameCount(); i++, samples += step )
+	for ( int i = 0; i < frames.GetFrameCount(); i++, samples += step )
 		* samples = ADSR::Tick();
 
 	return frames;
