+++ src/AudioCore/BiQuad.cpp
@@ -4,12 +4,12 @@
 NAMESPACE_AUDIO_BEGIN
 
 BiQuad::BiQuad() : Filter() {
-	b_.SetCount( 3, 0.0 );
-	a_.SetCount( 3, 0.0 );
-	b_[0] = 1.0;
-	a_[0] = 1.0;
-	inputs_.SetCount( 3, 1, 0.0 );
-	outputs_.SetCount( 3, 1, 0.0 );
+	b_.SetCount( 3, 0.0f );
+	a_.SetCount( 3, 0.0f );
+	b_[0] = 1.0f;
+	a_[0] = 1.0f;
+	inputs_.SetCount( 3, 1, 0.0f );
+	outputs_.SetCount( 3, 1, 0.0f );
 	Audio::AddSampleRateAlert( this );
 }
 
@@ -17,7 +17,7 @@ BiQuad::~BiQuad() {
 	Audio::RemoveSampleRateAlert( this );
 }
 
-void BiQuad::SetCoefficients( double b0, double b1, double b2, double a1, double a2, bool ClearState ) {
+void BiQuad::SetCoefficients( float b0, float b1, float b2, float a1, float a2, bool ClearState ) {
 	b_[0] = b0;
 	b_[1] = b1;
 	b_[2] = b2;
@@ -27,23 +27,23 @@ void BiQuad::SetCoefficients( double b0, double b1, double b2, double a1, double
 	if ( ClearState ) this->Clear();
 }
 
-void BiQuad::SampleRateChanged( double new_rate, double old_rate ) {
+void BiQuad::SampleRateChanged( float new_rate, float old_rate ) {
 	if ( !IgnoreSampleRateChange_ ) {
 		LOG("BiQuad::SampleRateChanged: you may need to recompute filter coefficients!");
 		HandleError( AudioError::WARNING );
 	}
 }
 
-void BiQuad::SetResonance( double frequency, double radius, bool normalize ) {
+void BiQuad::SetResonance( float frequency, float radius, bool normalize ) {
 	#if defined(flagDEBUG)
 
-	if ( frequency < 0.0 || frequency > 0.5 * Audio::GetSampleRate() ) {
+	if ( frequency < 0.0f || frequency > 0.5f * Audio::GetSampleRate() ) {
 		LOG("BiQuad::SetResonance: frequency argument (" << frequency << ") is out of range!");
 		HandleError( AudioError::WARNING );
 		return;
 	}
 
-	if ( radius < 0.0 || radius >= 1.0 ) {
+	if ( radius < 0.0f || radius >= 1.0f ) {
 		LOG("BiQuad::SetResonance: radius argument (" << radius << ") is out of range!");
 		HandleError( AudioError::WARNING );
 		return;
@@ -51,25 +51,25 @@ void BiQuad::SetResonance( double frequency, double radius, bool normalize ) {
 
 	#endif
 	a_[2] = radius * radius;
-	a_[1] = -2.0 * radius * cos( TWO_PI * frequency / Audio::GetSampleRate() );
+	a_[1] = -2.0f * radius * cos( TWO_PI * frequency / Audio::GetSampleRate() );
 
 	if ( normalize ) {
-		b_[0] = 0.5 - 0.5 * a_[2];
-		b_[1] = 0.0;
+		b_[0] = 0.5f - 0.5f * a_[2];
+		b_[1] = 0.0f;
 		b_[2] = -b_[0];
 	}
 }
 
-void BiQuad::SetNotch( double frequency, double radius ) {
+void BiQuad::SetNotch( float frequency, float radius ) {
 	#if defined(flagDEBUG)
 
-	if ( frequency < 0.0 || frequency > 0.5 * Audio::GetSampleRate() ) {
+	if ( frequency < 0.0f || frequency > 0.5f * Audio::GetSampleRate() ) {
 		LOG("BiQuad::SetNotch: frequency argument (" << frequency << ") is out of range!");
 		HandleError( AudioError::WARNING );
 		return;
 	}
 
-	if ( radius < 0.0 ) {
+	if ( radius < 0.0f ) {
 		LOG("BiQuad::SetNotch: radius argument (" << radius << ") is negative!");
 		HandleError( AudioError::WARNING );
 		return;
@@ -77,13 +77,13 @@ void BiQuad::SetNotch( double frequency, double radius ) {
 
 	#endif
 	b_[2] = radius * radius;
-	b_[1] = (double) - 2.0 * radius * cos( TWO_PI * (double) frequency / Audio::GetSampleRate() );
+	b_[1] = (float) - 2.0f * radius * cos( TWO_PI * (float) frequency / Audio::GetSampleRate() );
 }
 
 void BiQuad::SetEqualGainZeroes() {
-	b_[0] = 1.0;
-	b_[1] = 0.0;
-	b_[2] = -1.0;
+	b_[0] = 1.0f;
+	b_[1] = 0.0f;
+	b_[2] = -1.0f;
 }
 
 NAMESPACE_AUDIO_END
